+++ pancake2viper/src/ir/mod.rs
@@ -1,10 +1,8 @@
 mod expression;
 mod statement;
+mod toplevel;
 pub mod utils;
-
-pub mod expr_to_viper;
-pub mod heap_to_viper;
-pub mod stmt_to_viper;
 
 pub use expression::*;
 pub use statement::*;
+pub use toplevel::*;
