+++ pancake2viper/src/ir/toplevel.rs
@@ -1,5 +1,5 @@
 use super::Stmt;
-use crate::{pancake::Shape, translation::ToShape};
+use crate::{ir_to_viper::ViperEncodeCtx, shape::Shape, ToShape};
 
 #[derive(Debug, Clone)]
 pub struct FnDec {
@@ -9,7 +9,7 @@ pub struct FnDec {
 }
 
 impl<'a> ToShape<'a> for FnDec {
-    fn shape(&self, ctx: &crate::translation::ViperEncodeCtx<'a>) -> Shape {
+    fn shape(&self, ctx: &ViperEncodeCtx<'a>) -> Shape {
         // let shapes = FnDec::collect_returns(&self.body, ctx);
         // TODO: add check for types to be the same and non-empty
         // shapes[0].clone()
