+++ pancake2viper/src/ir_to_viper/mod.rs
@@ -2,3 +2,6 @@ pub mod expression;
 pub mod heap;
 pub mod statement;
 pub mod toplevel;
+mod utils;
+
+pub use utils::*;
