+++ pancake2viper/src/ir_to_viper/statement.rs
@@ -1,7 +1,10 @@
 use crate::ir;
+use crate::ir_to_viper::TranslationMode;
 use crate::utils::ViperUtils;
 
-use crate::translation::{context::ViperEncodeCtx, ToShape, ToViper, ToViperType};
+use crate::{ToShape, ToViper, ToViperType};
+
+use super::utils::ViperEncodeCtx;
 
 impl<'a> ToViper<'a, viper::Stmt<'a>> for ir::Stmt {
     fn to_viper(self, ctx: &mut ViperEncodeCtx<'a>) -> viper::Stmt<'a> {
@@ -38,12 +41,6 @@ impl<'a> ToViper<'a, viper::Stmt<'a>> for ir::Stmt {
         let seq = ast.seqn(&ctx.stack, &decls);
         ctx.stack.clear();
         seq
-    }
-}
-
-impl<'a> ToViper<'a, viper::Stmt<'a>> for ir::Annotation {
-    fn to_viper(self, ctx: &mut ViperEncodeCtx<'a>) -> viper::Stmt<'a> {
-        todo!()
     }
 }
 
@@ -102,7 +99,8 @@ impl<'a> ToViper<'a, viper::Stmt<'a>> for ir::While {
         body_seq.push(ast.label(&ctx.current_continue_label(), &[]));
         let body = ast.seqn(&body_seq, &[]);
 
-        ctx.stack.push(ast.while_stmt(cond, &[], body));
+        ctx.stack
+            .push(ast.while_stmt(cond, &body_ctx.invariants, body));
         ctx.stack.push(ast.label(&ctx.current_break_label(), &[]));
         let seq = ast.seqn(&ctx.stack, &decls);
         ctx.stack.clear();
@@ -161,8 +159,19 @@ impl<'a> ToViper<'a, viper::Stmt<'a>> for ir::Assign {
         let lhs_shape = ctx.get_type(&self.lhs);
         let name = ctx.mangler.mangle_var(&self.lhs);
         assert_eq!(lhs_shape, self.rhs.shape(ctx));
-        let var = ast.local_var(name, lhs_shape.to_viper_type(ctx));
-        let ass = ast.local_var_assign(var, self.rhs.to_viper(ctx));
+        let var = ast.new_var(name, lhs_shape.to_viper_type(ctx));
+        let ass = match self.rhs {
+            ir::Expr::MethodCall(call) => {
+                let mut args: Vec<viper::Expr> = call.args.to_viper(ctx);
+                args.insert(0, ctx.heap_var().1);
+                ast.method_call(
+                    &ctx.mangler.mangle_fn(&call.fname.label_to_viper()),
+                    &args,
+                    &[var.1],
+                )
+            }
+            other => ast.local_var_assign(var.1, other.to_viper(ctx)),
+        };
 
         let decls = ctx.pop_decls();
 
@@ -218,5 +227,32 @@ impl<'a> ToViper<'a, viper::Stmt<'a>> for ir::TailCall {
         );
         let goto = ast.goto(ctx.return_label());
         ast.seqn(&[call, goto], &[])
+    }
+}
+
+impl<'a> ToViper<'a, viper::Stmt<'a>> for ir::Annotation {
+    fn to_viper(self, ctx: &mut ViperEncodeCtx<'a>) -> viper::Stmt<'a> {
+        let ast = ctx.ast;
+        let no_pos = ast.no_position();
+
+        ctx.set_mode(self.typ.into());
+        let body = self.expr.to_viper(ctx);
+        ctx.set_mode(TranslationMode::Normal);
+        ctx.mangler.clear_annot_var();
+        use crate::ir::AnnotationType::*;
+        match self.typ {
+            Assertion => ast.assert(body, no_pos),
+            Inhale => ast.inhale(body, no_pos),
+            Exhale => ast.exhale(body, no_pos),
+            x => {
+                match x {
+                    Invariant => ctx.invariants.push(body),
+                    Precondition => ctx.pres.push(body),
+                    Postcondition => ctx.posts.push(body),
+                    _ => unreachable!(),
+                };
+                ast.comment("annotation pushed")
+            }
+        }
     }
 }
