+++ pancake2viper/src/lib.rs
@@ -2,12 +2,15 @@ pub mod annotation;
 pub mod ir;
 pub mod ir_to_viper;
 pub mod pancake;
+pub mod pancake_to_ir;
 pub mod parser;
-pub mod translation;
+pub mod shape;
+pub mod traits;
 pub mod utils;
 pub mod viper_prelude;
 
 #[cfg(test)]
 mod tests;
 
+pub use traits::*;
 pub use viper::Viper;
