+++ pancake2viper/src/pancake/mod.rs
@@ -1,12 +1,9 @@
 mod expression;
 mod parser;
-mod shape;
 mod statement;
-mod to_shape;
-mod top;
+mod toplevel;
 mod utils;
 
 pub use expression::*;
-pub use shape::Shape;
 pub use statement::*;
-pub use top::*;
+pub use toplevel::*;
