--- pancake2viper/src/pancake/top.rs
@@ -1,28 +1,0 @@
-use super::{shape::Shape, Stmt};
-use crate::translation::ToShape;
-
-#[derive(Debug, Clone)]
-pub struct FnDec {
-    pub fname: String,
-    pub args: Vec<Arg>,
-    pub body: Stmt,
-}
-
-impl<'a> ToShape<'a> for FnDec {
-    fn shape(&self, ctx: &crate::translation::ViperEncodeCtx<'a>) -> Shape {
-        let shapes = FnDec::collect_returns(&self.body, ctx);
-        // TODO: add check for types to be the same and non-empty
-        shapes[0].clone()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Arg {
-    pub name: String,
-    pub shape: Shape,
-}
-
-#[derive(Debug, Clone)]
-pub struct Program {
-    pub functions: Vec<FnDec>,
-}
