+++ pancake2viper/src/pancake/toplevel.rs
@@ -1,0 +1,20 @@
+use super::Stmt;
+use crate::{shape::Shape, ToShape};
+
+#[derive(Debug, Clone)]
+pub struct FnDec {
+    pub fname: String,
+    pub args: Vec<Arg>,
+    pub body: Stmt,
+}
+
+#[derive(Debug, Clone)]
+pub struct Arg {
+    pub name: String,
+    pub shape: Shape,
+}
+
+#[derive(Debug, Clone)]
+pub struct Program {
+    pub functions: Vec<FnDec>,
+}
